Specialize on inlined internal functions
Instead of specialising on the application of an internal function,
a function that is created as part of the compilation process,
we specialise on the inlined version.

This is needed when we specialise a recursive function on a
local function. If we didn't need this, we would specialise
on a different, new, function every time, because the compiler
doesn't consider the specialisation alpha-equivalent.

In some cases, this would prevent normalisation of functions
that are turned from local let-recursive functions to global
full recursive functios by GHC; because the Clash compiler
would hit a specialisation limit because none of the arguments
were alpha-equivalent.